Move StepList testing helpers to the StepList directory
