+++ packages/design-system/src/components/StepList/Step.test.jsx
@@ -1,4 +1,4 @@
-import { defaultStep, generateStep } from '../helpers/StepList';
+import { defaultStep, generateStep } from './__mocks__/generateStep';
 import React from 'react';
 import Step from './Step';
 import { shallow } from 'enzyme';
