+++ packages/design-system/src/components/StepList/__mocks__/generateStep.js
